requirements.txt
streamlit
sqlalchemy
sqlite3
python-dotenv
google-generative-ai
